tests/wpt/mozilla/tests/mozilla/mozbrowser/mozbrowsertitlechangedeagerly_event.html

<!doctype html>
<meta charset="utf-8">
<title>Check if title_changed is too eager (issue #10782)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body></body>
<script>
    async_test(function(t) {
        var iframe = document.createElement("iframe");
        iframe.mozbrowser = "true";
        iframe.src = "mozbrowsertitlechangedeagerly_event_iframe.html";
        iframe.addEventListener("mozbrowsertitlechange", t.step_func(e => {
            assert_equals(e.type, "mozbrowsertitlechange");
            assert_not_equals(e.detail, "Bar");
        }));
        iframe.addEventListener("load", t.step_func_done());
        document.body.appendChild(iframe);
    });
</script>
